requirements.txt
streamlit
pytesseract
Pillow
requests
dotenv
requests>=2.31.0
python-telegram-bot==13.15
python-dotenv
easyocr
